fix connection runtime error on windows

+++ python/xoscar/collective/tests/test_pygloo.py
@@ -28,9 +28,12 @@ system_name = platform.system()
 def worker_allgather(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    if os.path.exists(fileStore_path):
-        print(os.path)
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 2)
 
     if system_name == "Linux":
@@ -64,9 +67,6 @@ def worker_allgather(rank, fileStore_path):
 
 def test_allgather():
     temp_dir = "collective_allgather"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_allgather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_allgather, args=(1, temp_dir))
@@ -75,13 +75,19 @@ def test_allgather():
     process1.join()
     process2.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_allreduce(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 2)
 
     if system_name == "Linux":
@@ -115,9 +121,6 @@ def worker_allreduce(rank, fileStore_path):
 
 def test_allreduce():
     temp_dir = "collective_allreduce"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_allgather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_allgather, args=(1, temp_dir))
@@ -126,13 +129,19 @@ def test_allreduce():
     process1.join()
     process2.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_barrier(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 2)
 
     if system_name == "Linux":
@@ -167,9 +176,6 @@ def worker_barrier(rank, fileStore_path):
 
 def test_barrier():
     temp_dir = "collective_barrier"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_allgather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_allgather, args=(1, temp_dir))
@@ -178,13 +184,19 @@ def test_barrier():
     process1.join()
     process2.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_broadcast(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 2)
 
     if system_name == "Linux":
@@ -230,9 +242,6 @@ def worker_broadcast(rank, fileStore_path):
 
 def test_broadcast():
     temp_dir = "collective_broadcast"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_allgather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_allgather, args=(1, temp_dir))
@@ -241,13 +250,19 @@ def test_broadcast():
     process1.join()
     process2.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_gather(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 3)
 
     if system_name == "Linux":
@@ -287,9 +302,6 @@ def worker_gather(rank, fileStore_path):
 
 def test_gather():
     temp_dir = "collective_gather"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_gather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_gather, args=(1, temp_dir))
@@ -301,13 +313,19 @@ def test_gather():
     process2.join()
     process3.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_reduce_scatter(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 3)
 
     if system_name == "Linux":
@@ -357,9 +375,6 @@ def worker_reduce_scatter(rank, fileStore_path):
 @require_linux
 def test_reduce_scatter():
     temp_dir = "collective_reduce_scatter"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_gather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_gather, args=(1, temp_dir))
@@ -371,13 +386,19 @@ def test_reduce_scatter():
     process2.join()
     process3.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_reduce(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 3)
 
     if system_name == "Linux":
@@ -424,9 +445,6 @@ def worker_reduce(rank, fileStore_path):
 
 def test_reduce():
     temp_dir = "collective_reduce"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_gather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_gather, args=(1, temp_dir))
@@ -438,13 +456,19 @@ def test_reduce():
     process2.join()
     process3.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_scatter(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 2)
 
     if system_name == "Linux":
@@ -490,9 +514,6 @@ def worker_scatter(rank, fileStore_path):
 
 def test_scatter():
     temp_dir = "collective_scatter"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_allgather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_allgather, args=(1, temp_dir))
@@ -501,13 +522,19 @@ def test_scatter():
     process1.join()
     process2.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_send_recv(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 2)
 
     if system_name == "Linux":
@@ -555,9 +582,6 @@ def worker_send_recv(rank, fileStore_path):
 
 def test_send_recv():
     temp_dir = "collective_send_recv"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_allgather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_allgather, args=(1, temp_dir))
@@ -566,13 +590,19 @@ def test_send_recv():
     process1.join()
     process2.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
 
 
 def worker_all_to_all(rank, fileStore_path):
     from .. import xoscar_pygloo as xp
 
-    time.sleep(0.5)
+    if rank == 0:
+        if os.path.exists(fileStore_path):
+            shutil.rmtree(fileStore_path)
+        os.makedirs(fileStore_path)
+    else:
+        time.sleep(0.5)
     context = xp.rendezvous.Context(rank, 3)
 
     if system_name == "Linux":
@@ -604,9 +634,6 @@ def worker_all_to_all(rank, fileStore_path):
 
 def test_all_to_all():
     temp_dir = "collective_all_to_all"
-    if os.path.exists(temp_dir):
-        shutil.rmtree(temp_dir)
-    os.mkdir(temp_dir)
     process1 = mp.Process(target=worker_gather, args=(0, temp_dir))
     process1.start()
     process2 = mp.Process(target=worker_gather, args=(1, temp_dir))
@@ -618,4 +645,5 @@ def test_all_to_all():
     process2.join()
     process3.join()
 
-    shutil.rmtree(temp_dir)
+    if os.path.exists(temp_dir):
+        shutil.rmtree(temp_dir)
